fixed parseInt problem

+++ api/src/main/java/com/mlmstorenow/api/services/PaymentService.java
@@ -18,16 +18,15 @@ import com.mlmstorenow.api.models.User;
 @Service
 public class PaymentService {
 
-	
 	// creates static braintreegateway object
 	public static BraintreeGateway getBraintreeGateway() {
 
-			BraintreeGateway gateway = new BraintreeGateway(
-					Environment.SANDBOX, 
-					ConfigProperties.getConfigProp("mechant_id"),
-					ConfigProperties.getConfigProp("public_key"), 
-					ConfigProperties.getConfigProp("private_key"));
-			return gateway;
+		BraintreeGateway gateway = new BraintreeGateway(
+				Environment.SANDBOX,
+				ConfigProperties.getConfigProp("mechant_id"),
+				ConfigProperties.getConfigProp("public_key"),
+				ConfigProperties.getConfigProp("private_key"));
+		return gateway;
 	}
 
 	// pass clientToken to your front-end
@@ -41,7 +40,7 @@ public class PaymentService {
 	// Transaction processor
 	public String processTransaction(String chargeAmount, String nonce, String deviceData) {
 
-		BigDecimal bd = BigDecimal.valueOf(Integer.parseInt(chargeAmount));
+		BigDecimal bd = BigDecimal.valueOf(Float.parseFloat(chargeAmount));
 		TransactionRequest request = new TransactionRequest().amount(bd)
 				.paymentMethodNonce(nonce).deviceData(deviceData).options().submitForSettlement(true).done();
